Move registration tests into dedicated test suite
The suite will be changed in a separate commit to test not via `CommandBus` instance.
